Add another impl Trait allowedness parameter
In order to deny impl Trait in return position inside trait definitions,
add another switch to lower_fn_decl and lower_method_sig to indicate if
an existential impl Trait is allowed in that position.

This causes an E0195, lifetimes do not match error in
compile-fail/impl-trait/disallowed.rs. This is because everything is
collected in the impl, but not in the trait, as such the impl has
lifetime parameters in the HIR and later while the trait does not.  The
task now is to create a way to figure out if there is a way to determine
if an ImplItem is part of a trait impl just from the inputs to
lower_impl_item